+++ mldaikon/proxy_wrapper/proxy.py
@@ -1,40 +1,89 @@
 import inspect
 import json
-import linecache
 import logging
 import os
 import threading
+import torch.nn.parameter
+import torch
+import torch.nn.functional as F
+
+class FWrapper:
+    def __getattr__(self, name):
+        orig_attr = getattr(F, name)
+
+        if callable(orig_attr):
+            def wrapper(*args, **kwargs):
+                # unproxy the arguments
+                args = tuple(arg._obj if type(arg) is Proxy else arg for arg in args)
+                kwargs = {k: v._obj if type(v) is Proxy else v for k, v in kwargs.items()}
+                result = orig_attr(*args, **kwargs)
+                return proxy_handler(result, logdir=self.logdir, log_level=self.log_level, var_name=orig_attr.__name__)
+
+            return wrapper
+        else:
+            return orig_attr
+
+# Replace F with the wrapper
+F = FWrapper()
+
 import time
 from typing import (
+    Union,
+    Tuple,
     Any,
     Callable,
-    Dict,
     Iterator,
-    List,
-    Mapping,
-    Optional,
     Set,
-    Tuple,
-    TypeVar,
-    Union,
+    Optional,
     overload,
+    TypeVar,
+    Mapping,
+    Dict,
+    List,
+    Generator,
 )
-
-import torch
-import torch.nn.parameter
-
-import mldaikon.proxy_wrapper.proxy_methods as proxy_methods
-from mldaikon.proxy_wrapper.config import debug_mode, proxy_log_dir, proxy_update_limit
-from mldaikon.proxy_wrapper.dumper import dump_attributes, dump_meta_vars
-from mldaikon.proxy_wrapper.dumper import json_dumper as dumper
-from mldaikon.proxy_wrapper.dumper import torch_serialize
-from mldaikon.proxy_wrapper.utils import print_debug
+import types
 from mldaikon.utils import typename
+from mldaikon.proxy_wrapper.config import debug_mode, proxy_log_dir, proxy_update_limit, exclude_file_names
+from mldaikon.proxy_wrapper.dumper import json_dumper as dumper
+from mldaikon.proxy_wrapper.dumper import (
+    dump_tensor,
+    dump_attributes,
+    dump_meta_vars,
+    torch_serialize,
+)
+from mldaikon.proxy_wrapper.utils import print_debug
+import mldaikon.proxy_wrapper.proxy_methods as proxy_methods
+from mldaikon.proxy_wrapper.proxy_handler import handled_obj_type
+import linecache
 
 
 def get_line(filename, lineno):
     return linecache.getline(filename, lineno).strip()
 
+def proxy_handler(obj, logdir, log_level, var_name):
+    # if list or tuple, do the same thing for each element
+    if isinstance(obj, (list, tuple)):
+        for element in obj:
+            element = proxy_handler(element, logdir, log_level, var_name)
+    # # handle if type(obj) is <class 'generator'>
+    if isinstance(obj, types.GeneratorType):
+        import pdb; pdb.set_trace()
+        def generator_proxy_handler():
+            for element in obj:
+                yield proxy_handler(element, logdir, log_level, var_name)
+        obj = generator_proxy_handler()
+        
+    # if isinstance(obj, Iterator):
+    #     obj = (proxy_handler(element, logdir, log_level, var_name) for element in obj)
+    if typename(obj).startswith("torch.distributed"):
+        return obj
+    for obj_type in handled_obj_type:
+        if issubclass(type(obj), obj_type):
+            return Proxy(obj, logdir=logdir, log_level=log_level, var_name=var_name)
+    
+        
+    return obj
 
 class Proxy:
     proxy_dict = {}
@@ -80,16 +129,33 @@ class Proxy:
                 print_debug("logger_proxy: " + f"'{value}'")
 
     @staticmethod
-    def proxy_parameters(module, parent_name=""):
-        print(
-            "logger_proxy: "
-            + f"Proxying all parameters of '{parent_name + module.__class__.__name__}'"
-        )
+    def proxy_parameters(module, parent_name=''):
+        print("logger_proxy: " + f"Proxying all parameters of '{parent_name + module.__class__.__name__}'")
         for name, parameter in module.named_parameters():
             print("logger_proxy: " + f"Proxying parameter '{parent_name+name}'")
-            parameter = Proxy(parameter, var_name=parent_name + name)
-
-    def dump_to_trace(self, obj, status="update", dumped_frame_array=None):
+            parameter = Proxy(parameter, var_name = parent_name + name)
+            module._parameters[name] = parameter
+            
+    @staticmethod
+    def get_frame_array(frame):
+        frame_array = []
+        while frame:
+            if frame.f_code.co_filename in exclude_file_names:
+                frame = frame.f_back
+                continue
+            
+            # fetch the frame info
+            frame_array.append(
+                (
+                    frame.f_code.co_filename,
+                    frame.f_lineno,
+                    get_line(frame.f_code.co_filename, frame.f_lineno),
+                )
+            )
+            frame = frame.f_back
+        return frame_array
+    
+    def dump_to_trace(self, obj, status = "update", dumped_frame_array=None):
         if not issubclass(type(obj), torch.nn.Module):
             dumped_val = str(torch_serialize(obj))
             self.jsondumper.dump_json(
@@ -103,15 +169,8 @@ class Proxy:
                 dump_attributes(obj),
                 dumped_frame_array,
             )
-
-    def __init__(
-        self,
-        obj,
-        logdir="proxy_log.log",
-        log_level=logging.INFO,
-        is_root=False,
-        var_name="",
-    ):
+    
+    def __init__(self, obj, logdir="proxy_log.log", log_level=logging.INFO, is_root= False, var_name = ''):
         self.__dict__["process_id"] = os.getpid()
         self.__dict__["thread_id"] = threading.current_thread().ident
         self.__dict__["logdir"] = logdir
@@ -127,62 +186,24 @@ class Proxy:
                 "logger_proxy: "
                 + f"Object '{obj.__class__.__name__}' is already a proxy"
             )
-
+            
             # create a shallow copy of the object
             self._obj = obj._obj
             self.__dict__["dumped_varname_list"] = obj.__dict__["dumped_varname_list"]
-            self.__dict__["last_update_timestamp"] = obj.__dict__[
-                "last_update_timestamp"
-            ]
+            self.__dict__["last_update_timestamp"] = obj.__dict__["last_update_timestamp"]
             self.__dict__["is_proxied_obj"] = obj.__dict__["is_proxied_obj"]
             self.__dict__["is_root"] = obj.__dict__["is_root"]
             self.__dict__["var_name"] = obj.__dict__["var_name"]
             self.__dict__["logdir"] = obj.__dict__["logdir"]
             self.__dict__["log_level"] = obj.__dict__["log_level"]
-            self.__dict__["meta_vars"] = obj.__dict__["meta_vars"]
+            self.__dict__["meta_vars"] = obj.__dict__["meta_vars"]      
 
         else:
             frame = inspect.currentframe()
-            var_list = []
-            frame_array = []
-            while frame:
-                if frame.f_code.co_filename == __file__:
-                    frame = frame.f_back
-                else:
-                    # fetch the frame info
-                    frame_array.append(
-                        (
-                            frame.f_code.co_filename,
-                            frame.f_lineno,
-                            get_line(frame.f_code.co_filename, frame.f_lineno),
-                        )
-                    )
-
-                    # fetch the var_name from the stack_frame
-                    current_var_name = None
-                    for var_name, var_val in frame.f_locals.items():
-                        if var_val is obj or (
-                            isinstance(var_val, Proxy) and var_val._obj is obj
-                        ):
-                            current_var_name = var_name
-                            break
-                    if current_var_name is not None:
-                        print_debug(
-                            "logger_proxy: " + f"Variable name f{current_var_name}"
-                            "of the object is found in the current frame"
-                        )
-                        var_list.append(current_var_name)
-                    frame = frame.f_back
-
+            frame_array = self.get_frame_array(frame)
             dumped_frame_array = json.dumps(frame_array)
 
             # print_debug the variable name list of the object
-            if len(var_list) == 0:
-                print_debug("logger_proxy: " + f"Empty variable name list")
-                Proxy.empty_name_counts += 1
-            else:
-                print_debug("logger_proxy: " + f"Variable name list: {var_list}")
-                Proxy.non_empty_name_counts += 1
             print_debug(
                 "logger_proxy: " + f"Empty name counts: {Proxy.empty_name_counts}"
             )
@@ -191,50 +212,47 @@ class Proxy:
                 + f"Non-empty name counts: {Proxy.non_empty_name_counts}"
             )
 
-            # dumped var_list
-            # if len(var_list) == 0:
-            #     current_var_name_list = "None"
-            # else:
-            #     current_var_name_list = json.dumps(var_list)
             if self.__dict__["var_name"] is not None:
                 current_var_name_list = self.__dict__["var_name"]
             else:
-                current_var_name_list = ""
+                current_var_name_list = ''
             self.__dict__["dumped_varname_list"] = current_var_name_list
 
             # Ziming: here we still seperate the handling of tensor and other objects
             # however, despite the dumping logic these two are identical and could be merged
-            if self.__dict__["is_root"]:
+            if self.__dict__["is_root"] == True:
                 print("found the root object")
             if issubclass(type(obj), torch.nn.Module):
 
-                if self.__dict__["is_root"]:
+                if self.__dict__["is_root"] == True:
                     # proxy all of its parameters
 
                     self.proxy_parameters(obj)
                     for name, module in obj.named_children():
-                        setattr(obj, name, Proxy(module, var_name=name))
-                        self.proxy_parameters(module, name + ".")
+                        # setattr(obj, name, Proxy(module, var_name=name))
+                        proxy_module = Proxy(module, var_name=name)
+
+                        obj._modules[name] = proxy_module
+                        self.proxy_parameters(proxy_module, name+".")
                         # TODO: improve the nn.Module tracing logic, currently we only trace two levels of submodules
                         # We could try to enforce a blacklist of modules that we can't trace (contain low level functions)
                         if issubclass(type(module), torch.nn.Module):
                             for subname, submodule in module.named_children():
-                                self.proxy_parameters(
-                                    submodule, name + "." + subname + "."
-                                )
-                                setattr(
-                                    module,
-                                    subname,
-                                    Proxy(submodule, var_name=name + "." + subname),
-                                )
-
+                                proxy_submodule = Proxy(submodule, var_name=name+"."+subname)
+                                self.proxy_parameters(proxy_submodule, name+"."+subname+".")
+                                # setattr(module, subname, Proxy(submodule, var_name=name+"."+subname))
+                                
+                                module._modules[subname] = proxy_submodule
+                
                 else:
-                    if current_var_name_list == "":
+                    if current_var_name_list == '':
                         self.proxy_parameters(obj)
                     else:
-                        self.proxy_parameters(obj, current_var_name_list + ".")
-
-            if issubclass(type(obj), torch.Tensor):
+                        self.proxy_parameters(obj, current_var_name_list+".")    
+                               
+                                        
+            
+            if issubclass(type(obj),torch.Tensor):
                 # init tensor
                 tensor_shape = obj.shape.__str__()
                 current_var_name_list = current_var_name_list + tensor_shape
@@ -283,7 +301,7 @@ class Proxy:
 
                     Proxy.var_dict[current_var_name_list] = self
                 else:
-                    if type(obj) not in [int, float, str, bool] and obj is not None:
+                    if not type(obj) in [int, float, str, bool] and obj is not None:
                         print_debug(
                             "logger_proxy: "
                             + f"Object '{obj.__class__.__name__}' is already proxied"
@@ -318,53 +336,17 @@ class Proxy:
         self._obj.__class__.step(self, *args, **kwargs)
 
     def __call__(self, *args, **kwargs):
-        print_debug(
+        print(
             "logger_proxy: " + f"Go to __call__ for object '{self.__class__.__name__}'"
         )
         # only pass down the torch.nn.Module here
-        args = tuple(
-            (
-                arg._obj
-                if (
-                    type(arg) is Proxy
-                    and not issubclass(arg._obj, torch.nn.Module)
-                    and not issubclass(args._obj, torch.nn.parameter.Parameter)
-                )
-                else arg
-            )
-            for arg in args
-        )
-        kwargs = {
-            k: (
-                v._obj
-                if (
-                    type(v) is Proxy
-                    and not issubclass(v._obj, torch.nn.Module)
-                    and not issubclass(v._obj, torch.nn.parameter.Parameter)
-                )
-                else v
-            )
-            for k, v in kwargs.items()
-        }
-        print_debug(
-            "logger_proxy: "
-            + f"Calling '{self.__class__.__name__}' for obj: '{self.__dict__['var_name']}' (type '{typename(self._obj)}')"
-        )
+        args = tuple(arg._obj if (type(arg) is Proxy and not issubclass(arg._obj,torch.nn.Module) and not issubclass(args._obj,torch.nn.parameter.Parameter)) else arg for arg in args)
+        kwargs = {k: v._obj if (type(v) is Proxy  and not issubclass(v._obj,torch.nn.Module) and not issubclass(v._obj, torch.nn.parameter.Parameter)) else v for k, v in kwargs.items()}
+        print_debug("logger_proxy: " + f"Calling '{self.__class__.__name__}' for obj: '{self.__dict__['var_name']}' (type '{typename(self._obj)}')")
         result = self._obj(*args, **kwargs)
-        print_debug(
-            "logger_proxy: " + f"Result type of __call__ is '{typename(result)}'"
-        )
+        print_debug("logger_proxy: " + f"Result type of __call__ is '{typename(result)}'")
         # HACK: avoid proxying torch.distributed as we cannot handle ProcessGroup `in` ops in the get_group_rank & get_global_rank function
-        if typename(result).startswith("torch.distributed"):
-            return result
-        if issubclass(type(result), torch.nn.Module):
-            return Proxy(
-                result,
-                logdir=self.logdir,
-                log_level=self.log_level,
-                var_name=self.__dict__["var_name"],
-            )
-        return result
+        return proxy_handler(result, self.logdir, self.log_level, self.__dict__["var_name"])
 
     def __getattr__(self, name):
         print_debug("logger_proxy: " + f"Accessing attribute '{name}'")
@@ -377,59 +359,88 @@ class Proxy:
         # HACK: avoid proxying torch.distributed as we cannot handle ProcessGroup `in` ops in the get_group_rank & get_global_rank function
         if typename(attr).startswith("torch.distributed"):
             return attr
-
-        if self.__dict__["var_name"] == "":
+        
+        if self.__dict__["var_name"]=='':
             var_name = name
         else:
             var_name = self.__dict__["var_name"] + "." + name
         # if attr is a tensor or nn.Module, return a proxy object
-        if issubclass(
-            type(attr), torch.nn.parameter.Parameter
-        ):  # TODO: should double check if the torch.Tensor wrapping is effective here
-            return Proxy(
-                attr, logdir=self.logdir, log_level=self.log_level, var_name=var_name
-            )
-        if issubclass(type(attr), torch.nn.Module):
-            return Proxy(
-                attr, logdir=self.logdir, log_level=self.log_level, var_name=var_name
-            )
+        # if issubclass(type(attr), torch.nn.parameter.Parameter): # TODO: should double check if the torch.Tensor wrapping is effective here
+        #     # replace the attribute with a proxy object
+        #     proxy_attr = Proxy(attr, logdir=self.logdir, log_level=self.log_level, var_name = var_name)
+        #     return proxy_attr
 
+        # if issubclass(type(attr), torch.nn.Module):
+        #     proxy_attr =  Proxy(attr, logdir=self.logdir, log_level=self.log_level, var_name = var_name)
+        #     return proxy_attr
         # if attr is a bound method, return a wrapper function
         if callable(attr):
+            # if attr is a built-in function, return a wrapper function
+            # or function name starts with torch.nn.functional, return a wrapper function
+            # or function name starts with _, return a wrapper function
+            if type(attr) is types.BuiltinFunctionType or typename(attr).startswith("torch.nn.functional"):
+                def wrapper(*args, **kwargs):
+                    # import pdb; pdb.set_trace()
+                    print("logger_proxy: " + f"Calling torch.nn.functional method '{name}'")
+                    # unproxy the arguments recursively (take care of tuple and list obj for arg in args)
+                    args = list(args)
+                    # get rid of the last element if it is an empty dict
+                    if len(args) > 0 and type(args[-1]) is dict and len(args[-1]) == 0:
+                        args = args[:-1]
+                    i = -1
+                    # find the first proxy object in the args
+                    for i, arg in enumerate(args):
+                        if type(arg) in [tuple] and arg is not None and arg[0] is Proxy:
+                            break
+                    # get rid of the proxy object and before that, unproxy the arguments
+                    args = args[i+1:]
+                    
+                    def unproxy_arg(arg):
+                        
+                        if type(arg) is Proxy:
+                            return unproxy_arg(arg._obj)
+                        elif type(arg) in [list]:
+                            return [unproxy_arg(element) for element in arg]
+                        elif type(arg) in [tuple]:
+                            return tuple(unproxy_arg(element) for element in arg)
+                        else:
+                            return arg
+                    # args = [unproxy_arg(arg) for arg in args]
+                    if len(args) == 0:
+                        margs = tuple(args)
+                    else:
+                        margs = unproxy_arg(args[0])
+                    kwargs = {k: v._obj if type(v) is Proxy else v for k, v in kwargs.items()}
+                    
+                    result = attr(*margs, **kwargs)
+                    print_debug(f"Called method '{name}' with result {result}")
+                    return proxy_handler(result, self.logdir, self.log_level, var_name)
+                return wrapper
+            else:
+                attr = proxy_handler(attr, self.logdir, self.log_level, self.__dict__["var_name"])
+            # def method(*args, **kwargs):
+            #     result = Proxy(attr, var_name=self.__dict__["var_name"])(*args, **kwargs)
+            #     # print_debug(f"Called method '{name}' with result {result}")
+            #     # if result is not primitive, return a proxy object
+            #     return proxy_handler(result, self.logdir, self.log_level, var_name)
+        
 
-            def method(*args, **kwargs):
-                result = Proxy(attr, var_name=self.__dict__["var_name"])(
-                    *args, **kwargs
-                )
-                # print_debug(f"Called method '{name}' with result {result}")
-                # if result is not primitive, return a proxy object
-                if isinstance(result, torch.nn.parameter.Parameter):
-                    return Proxy(
-                        result,
-                        logdir=self.logdir,
-                        log_level=self.log_level,
-                        var_name=var_name,
-                    )
-                elif issubclass(type(result), torch.nn.Module):
-                    return Proxy(
-                        result,
-                        logdir=self.logdir,
-                        log_level=self.log_level,
-                        var_name=var_name,
-                    )
-                return result
-
-            return method
-        return attr
+            
+        return proxy_handler(attr, self.logdir, self.log_level, var_name)
 
     def __setattr__(self, name, value):
-        print_debug("logger_proxy: " + f"Setting attribute '{name}' to '{value}'")
+        
         if name == "_obj":
             self.__dict__[name] = value  # Set the attribute directly
         else:
-            self.dump_to_trace(value, "update")
-
-            if self.__dict__["var_name"] == "":
+            print("logger_proxy: " + f"Setting attribute '{name}' to '{value}'")
+            # dump frame array
+            frame = inspect.currentframe()
+            frame_array = self.get_frame_array(frame)
+            dumped_frame_array = json.dumps(frame_array)
+            self.dump_to_trace(self._obj, "update", dumped_frame_array)
+            
+            if self.__dict__["var_name"]=='':
                 var_name = name
             else:
                 var_name = self.__dict__["var_name"] + "." + name
@@ -438,12 +449,7 @@ class Proxy:
                 setattr(
                     self._obj,
                     name,
-                    Proxy(
-                        value,
-                        logdir=self.logdir,
-                        log_level=self.log_level,
-                        var_name=var_name,
-                    ),
+                    proxy_handler(value, logdir=self.logdir, log_level=self.log_level, var_name = var_name),
                 )
             else:
                 setattr(self._obj, name, value)