+++ mldaikon/proxy_wrapper/torch_proxy.py
@@ -1,11 +1,38 @@
+from typing import List, Tuple
 import torch
 import torch.distributed
+from torch.optim.optimizer import _get_fused_kernels_supported_devices, _get_foreach_kernels_supported_devices, _foreach_supported_types
 from torch._C._distributed_c10d import ReduceOp
 
 from .proxy import Proxy
 
 #################################################
 ###         Proxied Torch functions
+
+original_default_to_fused_or_foreach = torch.optim.optimizer._default_to_fused_or_foreach
+
+def _default_to_fused_or_foreach(params: List[torch.Tensor],
+                                 differentiable: bool,
+                                 use_fused: bool = False) -> Tuple[bool, bool]:
+    if torch.jit.is_scripting() or differentiable:
+        return False, False
+
+    fused_supported_devices = _get_fused_kernels_supported_devices()
+    foreach_supported_devices = _get_foreach_kernels_supported_devices()
+    fused = use_fused and all(
+        p is None or (type(p) in _foreach_supported_types or hasattr(p, 'is_proxied_obj') 
+                      and
+                      p.device.type in fused_supported_devices and
+                      torch.is_floating_point(p)) for p in params
+    )
+    foreach = not fused and all(
+        p is None or (type(p) in _foreach_supported_types  or hasattr(p, 'is_proxied_obj') 
+                      and
+                      p.device.type in foreach_supported_devices) for p in params
+    )
+    return fused, foreach
+
+torch.optim.optimizer._default_to_fused_or_foreach = _default_to_fused_or_foreach
 
 # Save the original broadcast function
 original_broadcast = torch.distributed.broadcast